code refactor for bench testing
