+++ src/ckb/tests/mod.rs
@@ -1,4 +1,6 @@
-pub mod actor;
+#[cfg(not(any(feature = "bench")))]
+mod actor;
+#[cfg(not(any(feature = "bench")))]
 mod config;
 
 pub mod test_utils;
