+++ src/fiber/channel.rs
@@ -1,4 +1,4 @@
-#[cfg(debug_assertions)]
+#[cfg(any(debug_assertions, feature = "bench"))]
 use crate::fiber::network::DebugEvent;
 use crate::fiber::network::PaymentCustomRecords;
 use crate::fiber::types::BroadcastMessageWithTimestamp;
@@ -101,9 +101,9 @@ pub const INITIAL_COMMITMENT_NUMBER: u64 = 0;
 const RETRYABLE_TLC_OPS_INTERVAL: Duration = Duration::from_millis(1000);
 
 // if a important TLC operation is not acked in 30 seconds, we will try to disconnect the peer.
-#[cfg(not(test))]
+#[cfg(not(any(test, feature = "bench")))]
 pub const PEER_CHANNEL_RESPONSE_TIMEOUT: u64 = 30 * 1000;
-#[cfg(test)]
+#[cfg(any(test, feature = "bench"))]
 pub const PEER_CHANNEL_RESPONSE_TIMEOUT: u64 = 10 * 1000;
 
 #[derive(Debug)]
@@ -150,7 +150,7 @@ pub enum ChannelCommand {
     BroadcastChannelUpdate(),
     Update(UpdateCommand, RpcReplyPort<Result<(), String>>),
     ForwardTlcResult(ForwardTlcResult),
-    #[cfg(test)]
+    #[cfg(any(test, feature = "bench"))]
     ReloadState(ReloadParams),
 }
 
@@ -165,19 +165,19 @@ impl Display for ChannelCommand {
             ChannelCommand::BroadcastChannelUpdate() => write!(f, "BroadcastChannelUpdate"),
             ChannelCommand::Update(_, _) => write!(f, "Update"),
             ChannelCommand::ForwardTlcResult(res) => write!(f, "ForwardTlcResult [{:?}]", res),
-            #[cfg(test)]
+            #[cfg(any(test, feature = "bench"))]
             ChannelCommand::ReloadState(_) => write!(f, "ReloadState"),
         }
     }
 }
 
-#[cfg(test)]
+#[cfg(any(test, feature = "bench"))]
 #[derive(Debug)]
 pub struct ReloadParams {
     pub notify_changes: bool,
 }
 
-#[cfg(test)]
+#[cfg(any(test, feature = "bench"))]
 impl Default for ReloadParams {
     fn default() -> Self {
         Self {
@@ -1938,7 +1938,7 @@ where
                     .await;
                 Ok(())
             }
-            #[cfg(test)]
+            #[cfg(any(test, feature = "bench"))]
             ChannelCommand::ReloadState(reload_params) => {
                 *state = self
                     .store
@@ -2862,7 +2862,6 @@ pub struct TlcState {
 }
 
 impl TlcState {
-    #[cfg(debug_assertions)]
     pub fn debug(&self) {
         for tlc in self.offered_tlcs.tlcs.iter() {
             debug!("offered_tlc: {:?}", tlc.log());
