+++ src/fiber/tests/mod.rs
@@ -8,6 +8,5 @@ mod path;
 mod payment;
 mod rpc;
 mod serde_utils;
-pub mod test_utils;
 mod tlc_op;
 mod types;
