+++ src/store/tests/store.rs
@@ -8,7 +8,6 @@ use crate::fiber::history::Direction;
 use crate::fiber::history::TimedResult;
 use crate::fiber::network::PaymentCustomRecords;
 use crate::fiber::network::SendPaymentData;
-use crate::fiber::tests::test_utils::*;
 use crate::fiber::types::*;
 use crate::gen_rand_fiber_private_key;
 use crate::gen_rand_fiber_public_key;
@@ -18,6 +17,7 @@ use crate::now_timestamp_as_millis_u64;
 use crate::store::store::deserialize_from;
 use crate::store::store::serialize_to_vec;
 use crate::store::Store;
+use crate::tests::test_utils::*;
 use crate::watchtower::*;
 use ckb_hash::new_blake2b;
 use ckb_types::packed::*;
